feat: move mock issuer to /credentials/issue route
- Add ed/credentials/issue/[documentType] page for issuing document
  credentials, matching the app's existing layout and theme
- Updated UserCredentials navigation to use new route so the sidebar
  Credentials tab highlights correctly
- Remove dold /mock-issuer route
